requirements.txt
Package            Version
------------------ ------------
aiohttp            3.9.5
aiosignal          1.3.1
async-timeout      4.0.3
attrs              23.2.0
bert-score         0.3.13
certifi            2024.7.4
charset-normalizer 3.3.2
contourpy          1.2.1
cycler             0.12.1
datasets           2.20.0
dill               0.3.8
filelock           3.15.4
fonttools          4.53.1
frozenlist         1.4.1
fsspec             2024.5.0
huggingface-hub    0.24.2
idna               3.7
kiwisolver         1.4.5
matplotlib         3.9.0
multidict          6.0.5
multiprocess       0.70.16
numpy              1.24.3
packaging          24.1
pandas             2.2.2
pillow             10.4.0
pip                24.2
pyarrow            17.0.0
pyarrow-hotfix     0.6
pyparsing          3.1.2
python-dateutil    2.9.0.post0
pytz               2024.1
PyYAML             6.0.1
regex              2024.7.24
requests           2.32.3
safetensors        0.4.3
seaborn            0.13.2
setuptools         70.1.1
six                1.16.0
tokenizers         0.19.1
torch              1.12.1+cu116
torchaudio         0.12.1+cu116
torchvision        0.13.1+cu116
tqdm               4.66.4
transformers       4.43.3
typing_extensions  4.12.2
tzdata             2024.1
urllib3            2.2.2
wheel              0.43.0
xxhash             3.4.1
yarl               1.9.4
